Refactor: Use conftest for some fixtures in channel tests
