+++ apps/channels/tests/test_facebook_integration.py
@@ -8,10 +8,8 @@ from apps.channels.datamodels import TwilioMessage
 from apps.channels.models import ChannelPlatform
 from apps.channels.tasks import handle_twilio_message
 from apps.chat.channels import MESSAGE_TYPES
-from apps.service_providers.models import MessagingProviderType
 from apps.service_providers.speech_service import SynthesizedAudio
 from apps.utils.factories.channels import ExperimentChannelFactory
-from apps.utils.factories.service_provider_factories import MessagingProviderFactory
 
 from .message_examples import twilio_messages
 
@@ -23,13 +21,6 @@ def _twilio_whatsapp_channel(twilio_provider):
         messaging_provider=twilio_provider,
         experiment__team=twilio_provider.team,
         extra_data={"page_id": "14155238886"},
-    )
-
-
-@pytest.fixture()
-def twilio_provider(db):
-    return MessagingProviderFactory(
-        name="twilio", type=MessagingProviderType.twilio, config={"auth_token": "123", "account_sid": "123"}
     )
